refactor(trusty-common): split kg_redb into submodules under 500-SLOC cap (#607)
Splits the 1400-SLOC `kg_redb.rs` monolith into six focused files under
`kg_redb/`:

- `types.rs` — shared constants, KgDbState, db_cache, BatchWriteOp,
  BatchOpResult, and pure helper functions
- `store.rs` — KgStoreRedb struct, open(), is_read_only(), db(), check_writable()
- `write_ops.rs` — assert, retract, upsert_drawer, delete_drawer,
  delete_by_subject, and in-transaction batch helpers
- `read_ops.rs` — query_active, list_subjects*, list_active, count_active_triples,
  checkpoint, load_drawers, load_drawer_ids, dump_all_triples
- `import.rs` — import_all and apply_batch
- `tests.rs` — all 20 unit tests (classified as test file, 1500-SLOC cap)
- `mod.rs` — thin re-export facade (~30 SLOC)

All 20 kg_redb tests pass, all 312 trusty-common lib tests pass, clippy
-D warnings clean, fmt clean. Drops kg_redb.rs from .line-cap-allowlist.tsv.
